Update to make the code more testable.
This invovled:
* Pulled out the output code (which incidentatlly should prevent file
    clobering in the future).
* Converted the conversion class back into a function. There might be reasons
    to make it a class again in the future, but it seemed that the only reason
    it was a class in the first place was for the input output which have been
    split out.
* Pulled out the output code into a separate section. Also moved the csv code
    into there. There may be reasons to create other types of output, although
    that's the entire point of the script right now.

Haven't actually added testing, but it should be easy enough to do for the
convert function, the input output could really be converted to use file
like objects rather that open files directly, but for the current scope of
the prgram this is probably enough.

Lastly this should really use argparse instead of the simplistic way arguments
are currently being split.