Added a PySide gui to the program.
Moved the logic into a new module and updated the console version and gui
to call that.

TODO:
- "main" and "window" are not very descriptive names
- There is shared logic for checking input out paths which should be
    standardized in a central location.
- Tests should be written.

This actually ended up being a lot of work for a script that I hope to one day
scrap... Although the gui is generic enough that I'm sure I can re-use it later
as a base for other scripts.