Add requirements.txt file.
Not that there's much there, and installing pyside requires a bunch of other
requirments on an average system, but hey.

I'm not sure if it's worth looking into anything else, I haven't acutlaly used
qt for a while, but for now it works. I assume other options have simliar
requirements, so I guess no rush.